Add save method (for syncing)
The save endpoint allows you to overwrite the whole calendar
with one that you pass through JSON.

Its there because I was worried that some requests might get lost
in transit and then the frontend state  wouldn't be in sync with the
backend state. So this can be used when you need to make sure that
what you're seeing in the frontend is reflected on the backend.

The other option was to request the whole calendar after any change
on the frontend, but I felt that would be too unresponsive.